clean build warnings

+++ gbmcmc/src/GalacticBinaryFStatistic.c
@@ -233,7 +233,7 @@ void get_filters(struct Orbit *orbit, struct Data *data, int filter_id, struct F
         // map to conventions for waveform generator
         params[3]=log(params[3]);
         params[4]=cos(params[4]);
-        galactic_binary(orbit, data->format, data->T, data->t0[0], params, data->NP, F_filter->A1_fX, F_filter->A1_fY, F_filter->A1_fZ, F_filter->A1_fA, F_filter->A1_fE, M_filter, 2);
+        galactic_binary(orbit, data->format, data->T, data->t0[0], params, data->NP, F_filter->A1_fX, F_filter->A1_fY, F_filter->A1_fZ, F_filter->A1_fA, F_filter->A1_fE, (int)M_filter, 2);
         
     } else if (filter_id == 2){
         
@@ -252,7 +252,7 @@ void get_filters(struct Orbit *orbit, struct Data *data, int filter_id, struct F
         // map to conventions for waveform generator
         params[3]=log(params[3]);
         params[4]=cos(params[4]);
-        galactic_binary(orbit, data->format, data->T, data->t0[0], params, data->NP, F_filter->A2_fX, F_filter->A2_fY, F_filter->A2_fZ, F_filter->A2_fA, F_filter->A2_fE, M_filter, 2);
+        galactic_binary(orbit, data->format, data->T, data->t0[0], params, data->NP, F_filter->A2_fX, F_filter->A2_fY, F_filter->A2_fZ, F_filter->A2_fA, F_filter->A2_fE, (int)M_filter, 2);
         
     } else if (filter_id == 3){
         
@@ -271,7 +271,7 @@ void get_filters(struct Orbit *orbit, struct Data *data, int filter_id, struct F
         // map to conventions for waveform generator
         params[3]=log(params[3]);
         params[4]=cos(params[4]);
-        galactic_binary(orbit, data->format, data->T, data->t0[0], params, data->NP, F_filter->A3_fX, F_filter->A3_fY, F_filter->A3_fZ, F_filter->A3_fA, F_filter->A3_fE, M_filter, 2);
+        galactic_binary(orbit, data->format, data->T, data->t0[0], params, data->NP, F_filter->A3_fX, F_filter->A3_fY, F_filter->A3_fZ, F_filter->A3_fA, F_filter->A3_fE, (int)M_filter, 2);
         
     } else {
         
@@ -290,7 +290,7 @@ void get_filters(struct Orbit *orbit, struct Data *data, int filter_id, struct F
         // map to conventions for waveform generator
         params[3]=log(params[3]);
         params[4]=cos(params[4]);
-        galactic_binary(orbit, data->format, data->T, data->t0[0], params, 9, F_filter->A4_fX, F_filter->A4_fY, F_filter->A4_fZ, F_filter->A4_fA, F_filter->A4_fE, M_filter, 2);
+        galactic_binary(orbit, data->format, data->T, data->t0[0], params, 9, F_filter->A4_fX, F_filter->A4_fY, F_filter->A4_fZ, F_filter->A4_fA, F_filter->A4_fE, (int)M_filter, 2);
     }
     
     free(params);
@@ -369,7 +369,7 @@ void get_M(struct Filter *F_filter, double **M_inv_X, double **M_inv_AE, struct 
     
     for (i=0; i<M; i++)
     {
-        k = (q + i - M_filter/2);
+        k = ((int)q + i - (int)M_filter/2);
         
         if(k>0 && k<data->N)
         {
